+++ npm-debug.log
@@ -1,0 +1,35 @@
+0 info it worked if it ends with ok
+1 verbose cli [ 'node', '/usr/local/bin/npm', 'start' ]
+2 info using npm@2.10.1
+3 info using node@v0.12.4
+4 verbose run-script [ 'prestart', 'start', 'poststart' ]
+5 info prestart brandle@0.0.0
+6 info start brandle@0.0.0
+7 verbose unsafe-perm in lifecycle true
+8 info brandle@0.0.0 Failed to exec start script
+9 verbose stack Error: brandle@0.0.0 start: `node ./bin/www`
+9 verbose stack Exit status 1
+9 verbose stack     at EventEmitter.<anonymous> (/usr/local/lib/node_modules/npm/lib/utils/lifecycle.js:213:16)
+9 verbose stack     at EventEmitter.emit (events.js:110:17)
+9 verbose stack     at ChildProcess.<anonymous> (/usr/local/lib/node_modules/npm/lib/utils/spawn.js:24:14)
+9 verbose stack     at ChildProcess.emit (events.js:110:17)
+9 verbose stack     at maybeClose (child_process.js:1015:16)
+9 verbose stack     at Process.ChildProcess._handle.onexit (child_process.js:1087:5)
+10 verbose pkgid brandle@0.0.0
+11 verbose cwd /Users/Scott/Desktop/School/CPSC/448/test
+12 error Darwin 14.3.0
+13 error argv "node" "/usr/local/bin/npm" "start"
+14 error node v0.12.4
+15 error npm  v2.10.1
+16 error code ELIFECYCLE
+17 error brandle@0.0.0 start: `node ./bin/www`
+17 error Exit status 1
+18 error Failed at the brandle@0.0.0 start script 'node ./bin/www'.
+18 error This is most likely a problem with the brandle package,
+18 error not with npm itself.
+18 error Tell the author that this fails on your system:
+18 error     node ./bin/www
+18 error You can get their info via:
+18 error     npm owner ls brandle
+18 error There is likely additional logging output above.
+19 verbose exit [ 1, true ]
