+++ public/javascripts/controllers/authController.js
@@ -1,0 +1,24 @@
+angular.module('brandle').controller('AuthCtrl', [
+    '$scope',
+    '$state',
+    'auth',
+    function($scope, $state, auth){
+      $scope.user = {};
+
+      $scope.register = function(){
+        auth.register($scope.user).error(function(error){
+          $scope.error = error;
+        }).then(function(){
+          $state.go('home');
+        });
+      };
+
+      $scope.logIn = function(){
+        auth.logIn($scope.user).error(function(error){
+          $scope.error = error;
+        }).then(function(){
+          $state.go('home');
+        });
+      };
+    }
+]);
