+++ public/javascripts/controllers/navController.js
@@ -1,0 +1,9 @@
+angular.module('brandle').controller('NavCtrl', [
+    '$scope',
+    'auth',
+    function($scope, auth){
+      $scope.isLoggedIn = auth.isLoggedIn;
+      $scope.currentUser = auth.currentUser;
+      $scope.logOut = auth.logOut;
+    }
+]);
