+++ views/index.ejs
@@ -7,7 +7,7 @@
     <script src="/javascripts/brandleApp.js"></script>
     <style> .glyphicon-thumbs-up { cursor:pointer } </style>
   </head>
-  <body ng-app="brandle" >
+  <body ng-app="brandle">
     <nav class="navbar navbar-default pull-right" ng-controller="NavCtrl">
       <ul class="nav navbar-nav">
         <li ng-show="isLoggedIn()"><a>{{ currentUser() }}</a></li>
@@ -19,150 +19,57 @@
     <div class="row">
       <div class="col-md-6 col-md-offset-3">
         <ui-view></ui-view>
+
+        <!--HOME-->
         <script type="text/ng-template" id="/home.html">
           <div class="page-header">
             <h1>Brandle</h1>
           </div>
-	  <span ng-show="post.author">
-	    posted by <a>{{post.author}}</a> |
-          </span>
-          <div ng-repeat="post in posts | orderBy:'-upvotes'">
-            <span class="glyphicon glyphicon-thumbs-up"
-              ng-click="incrementUpvotes(post)"></span>
-              {{post.upvotes}}
-            <span style="font-size:20px; margin-left:10px;">
-              <a ng-show="post.link" href="{{post.link}}">
-                {{post.title}}
-              </a>
-              <span ng-hide="post.link">
-                {{post.title}}
-              </span>
-              <span>
-	        <a href="#/posts/{{post._id}}">Comments</a>
-              </span>
-            </span>
-          </div>
-	  <div ng-hide="isLoggedIn()">
-	    <h3>You need to <a href="/#/login">Log In</a> or <a href="/#/register">Register</a> before you can add a post.</h3>
-	  </div>
-          
-          <form ng-submit="addPost()"
-	    ng-show="isLoggedIn()"
-            style="margin-top:30px;">
-            <h3>Add a new post</h3>
-          
-            <div class="form-group">
-              <input type="text"
-              class="form-control"
-              placeholder="Title"
-              ng-model="title"></input>
-            </div>
-            <div class="form-group">
-              <input type="text"
-              class="form-control"
-              placeholder="Link"
-              ng-model="link"></input>
-            </div>
-            <button type="submit" class="btn btn-primary">Post</button>
-	    sts/5590d9ed639ecb0360caad7a
-          </form>
-          
         </script>
-      </div>
-    </div>
-    <script type="text/ng-template" id="/posts.html">
-      <div class="page-header">
-        <h3>
-          <a ng-show="post.link" href="{{post.link}}">
-            {{post.title}}
-          </a>
-          <span ng-hide="post.link">
-            {{post.title}}
-          </span>
-        </h3>
-      </div>
-      
-      <div ng-repeat="comment in post.comments | orderBy:'-upvotes'">
-        <span class="glyphicon glyphicon-thumbs-up"
-          ng-click="incrementUpvotes(comment)">
-	</span>
-        {{comment.upvotes}} - by {{comment.author}}
-        <span style="font-size:20px; margin-left:10px;">
-          {{comment.body}}
-        </span>
-      </div>
-      <div ng-hide="isLoggedIn()">
-        <h3>You need to <a href="/#/login">Log In</a> or <a href="/#/register">Register</a> before you can comment.</h3>
-      </div>
-      <form ng-submit="addComment()"
-        ng-show="isLoggedIn()"
-	style="margin-top:30px;">
-	<h3>Add a new comment</h3>
-	 
-	<div class="form-group">
-	  <input type="text"
-	  class="form-control"
-	  placeholder="Comment"
-	  ng-model="body"></input>
-	</div>
-	<button type="submit" class="btn btn-primary">Post</button>
-      </form>
-    </script>
-    <script type="text/ng-template" id="/register.html">
-      <div class="page-header">
-	<h1>Brandle</h1>
-      </div>
 
-      <div ng-show="error" class="alert alert-danger row">
-	<span>{{ error.message }}</span>
-      </div>
+        <!--REGISTER-->
+        <script type="text/ng-template" id="/register.html">
+          <div class="page-header">
+            <h1>Brandle</h1>
+          </div>
 
-      <form ng-submit="register()"
-	style="margin-top:30px;">
-	<h3>Register</h3>
+          <div ng-show="error" class="alert alert-danger row">
+            <span>{{ error.message }}</span>
+          </div>
 
-	<div class="form-group">
-	  <input type="text"
-	  class="form-control"
-	  placeholder="Username"
-	  ng-model="user.username"></input>
-	</div>
-	<div class="form-group">
-	  <input type="password"
-	  class="form-control"
-	  placeholder="Password"
-	  ng-model="user.password"></input>
-	</div>
-	<button type="submit" class="btn btn-primary">Register</button>
-      </form>
-    </script>
-    <script type="text/ng-template" id="/login.html">
-      <div class="page-header">
-	<h1>Brandle</h1>
-      </div>
+          <form ng-submit="register()" style="margin-top:30px;">
+            <h3>Register</h3>
 
-      <div ng-show="error" class="alert alert-danger row">
-	<span>{{ error.message }}</span>
-      </div>
+          	<div class="form-group">
+          	  <input type="text" class="form-control" placeholder="Username" ng-model="user.username"></input>
+          	</div>
+          	<div class="form-group">
+          	  <input type="password" class="form-control" placeholder="Password" g-model="user.password"></input>
+          	</div>
+          	<button type="submit" class="btn btn-primary">Register</button>
+          </form>
+        </script>
 
-      <form ng-submit="logIn()"
-	style="margin-top:30px;">
-	<h3>Log In</h3>
+        <!--LOGIN-->
+        <script type="text/ng-template" id="/login.html">
+          <div class="page-header">
+            <h1>Brandle</h1>
+          </div>
 
-	<div class="form-group">
-	  <input type="text"
-	  class="form-control"
-	  placeholder="Username"
-	  ng-model="user.username"></input>
-	</div>
-	<div class="form-group">
-	  <input type="password"
-	  class="form-control"
-	  placeholder="Password"
-	  ng-model="user.password"></input>
-	</div>
-	<button type="submit" class="btn btn-primary">Log In</button>
-      </form>
-    </script>
+          <div ng-show="error" class="alert alert-danger row">
+            <span>{{ error.message }}</span>
+          </div>
+
+          <form ng-submit="logIn()" style="margin-top:30px;">
+            <h3>Log In</h3>
+            <div class="form-group">
+              <input type="text" class="form-control" placeholder="Username" ng-model="user.username"></input>
+            </div>
+          	<div class="form-group">
+          	  <input type="password" class="form-control" placeholder="Password" ng-model="user.password"></input>
+          	</div>
+            <button type="submit" class="btn btn-primary">Log In</button>
+          </form>
+        </script>
   </body>
 </html>
